Serving assets out of ./dist instead of moving files around
With the changes in 1a4c124, we now have more than just an index.html file in
the doc root. So we could either move all files in ./dist out into the root
dir, or we could serve up everything in ./dist. The latter sounds a bit more
flexible and less prone to errors.